# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virta API",
        "description": "API for managing companies and their stations.",
        "contact": {
            "name": "Michael Govern",
            "email": "michaelgovern0826@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/company": {
            "get": {
                "summary": "Get a list of companies",
                "description": "Returns a list of all companies.",
                "operationId": "3e0bffb951b3f31f83aadc004638122b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "parent_company_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new company",
                "description": "Create a new company record.",
                "operationId": "faf1545d2af89079ca55c4f9ed6ea302",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_company_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_company_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/company/{company}": {
            "get": {
                "summary": "Get a specific company",
                "description": "Returns details of a specific company.",
                "operationId": "339150f4b4353645e1ef425708b4f2f7",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of the company to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_company_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            },
            "put": {
                "summary": "Update a company",
                "description": "Update a company record.",
                "operationId": "3f0d99f5f2607c770763609c3d620829",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of the company to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parent_company_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "summary": "Delete a company",
                "description": "Delete a company record.",
                "operationId": "066ab74a6464b2ea1b3551447b803bc1",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of the company to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Company deleted successfully"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
        "/api/child-stations/{company_id}": {
            "get": {
                "summary": "Get stations owned by a company and its children",
                "description": "Returns a list of stations owned by a company and its children.",
                "operationId": "170bfad14cbc1b46ed3e70ca45df0984",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "ID of the company to retrieve stations for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "company_id": {
                                                "type": "integer"
                                            },
                                            "address": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
...
    }
}